Moved Interface to another file
